// src/generated/types.ts
export interface paths {
// ...
    "/wallets": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Wallets
         * @description Fetch a list of wallets.
         */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successfully retrieved the list of walllets. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @example true */
                            success?: boolean;
                            data?: {
                                wallets?: components["schemas"]["Wallet"][];
                            };
                        };
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        put?: never;
        /**
         * Create Wallet
         * @description Create a new wallet.
         */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            /** @description The details required to create a wallet. */
            requestBody: {
                content: {
                    "application/json": components["schemas"]["PostWalletRequest"];
                };
            };
            responses: {
                /** @description Successfully created wallet. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["DataResponse"];
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/wallets/{wallet_id}/statements": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Statements
         * @description Fetch a list of generated wallet statements.
         */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    wallet_id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successfully retrieved the list of wallet statements. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @example true */
                            success?: boolean;
                            data?: {
                                statements?: components["schemas"]["WalletStatement"][];
                            };
                        };
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/wallets/{wallet_id}/transactions": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Transactions
         * @description Fetch a list of wallet transactions.
         */
        get: {
            parameters: {
                query?: {
                    balance_id?: string;
                    status?: "processing" | "settled" | "requires-approval" | "declined" | "failed" | "cancelled";
                    type?: "debit" | "credit";
                    method?: "P2P" | "A2A" | "TopUp" | "Payout" | "EcomPay";
                    limit?: number;
                };
                header?: never;
                path: {
                    wallet_id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successfully retrieved the list of wallet transactions. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @example true */
                            success?: boolean;
                            data?: {
                                transactions?: (components["schemas"]["WalletOutboundTransaction"] | components["schemas"]["WalletInboundTransaction"] | components["schemas"]["WalletPayTransaction"])[];
                            };
                        };
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/wallets/{wallet_id}/balances": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Balances
         * @description Fetch a list of wallet balances.
         */
        get: {
            parameters: {
                query?: {
                    type?: components["schemas"]["WalletBalanceType"];
                    currency?: string;
                };
                header?: never;
                path: {
                    wallet_id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successfully retrieved the list of wallet balances. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @example true */
                            success?: boolean;
                            data?: {
                                balances?: components["schemas"]["WalletBalance"][];
                            };
                        };
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        put?: never;
        /**
         * Open Balance
         * @description Open a new wallet balance.
         */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    wallet_id: string;
                };
                cookie?: never;
            };
            /** @description The details required to create a balance. */
            requestBody: {
                content: {
                    "application/json": components["schemas"]["PostWalletBalanceRequest"];
                };
            };
            responses: {
                /** @description Successfully created wallet balance. */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["DataResponse"];
                    };
                };
                /** @description Client error. */
                "4XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
                /** @description Server error. */
                "5XX": {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["ErrorResponse"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
// ...
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}